# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mansory API",
        "description": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/brands": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Bütün markaları gətir",
                "operationId": "130f2bd423b6fe2355253e2c39878d99",
                "responses": {
                    "200": {
                        "description": "Uğurlu"
                    }
                }
            },
            "post": {
                "tags": [
                    "Brands"
                ],
                "summary": "Yeni marka yarat",
                "operationId": "283de011874bd2a1ada6efafb676e83f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Mercedes-Benz"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Luxury vehicles"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Yaradıldı"
                    },
                    "422": {
                        "description": "Validation xətası"
                    }
                }
            }
        },
        "/api/brands/{id}": {
            "get": {
                "tags": [
                    "Brands"
                ],
                "summary": "Bir markanı gətir",
                "operationId": "2086e16f9b1a70aea714677ef42808a7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Uğurlu"
                    },
                    "404": {
                        "description": "Tapılmadı"
                    }
                }
            },
            "put": {
                "tags": [
                    "Brands"
                ],
                "summary": "Markanı yenilə",
                "operationId": "40baf1fc4e11f1ccc369eb0b36b56f54",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Yeniləndi"
                    },
                    "404": {
                        "description": "Tapılmadı"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Brands"
                ],
                "summary": "Markanı sil",
                "operationId": "b3f38ad7a979e5c78ac4e95ceb640c7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Silindi"
                    },
                    "404": {
                        "description": "Tapılmadı"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}